Separate real, dummy implementations of syscall module.
